# Track local frames incrementally during execution