@RalfJung 2022-11-04
So those benchmarks are without https://github.com/rust-lang/miri/pull/2646?