@saethlin 2022-11-04
> So those benchmarks are without #2646?

Yes. That PR adds extra work to the data race machinery in order to store the spans which has its own very small regression, so I wanted to keep the effects separate.