@RalfJung 2022-11-20
LGTM, what are the benchmarks saying? How could the cache be invalidated less aggressively than on each pop?