@saethlin 2022-11-21
> How could the cache be invalidated less aggressively than on each pop?

I force-pushed after making this comment. Previously I was invalidating on every pop, not just pops of local frames. Most frames we create are in the standard library.