@saethlin 2022-11-21
> what are the benchmarks saying?

I updated the PR description. We have the same marginal improvement.